+++ lib/rails_ai_context/serializers/copilot_instructions_serializer.rb
@@ -45,7 +45,7 @@ module RailsAiContext
 
         schema = context[:schema]
         if SectionGuard.usable?(schema)
-          lines << "- Database: #{schema[:adapter]} - #{count_phrase(schema[:total_tables], "table")}"
+          lines << "- Database: #{database_adapter_label(schema)} - #{count_phrase(schema[:total_tables], "table")}"
         end
 
         models = context[:models]